Move BuildRunResultRegisterer and StepActivator to their own go files
