+++ cli/run_util.go
@@ -7,27 +7,21 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"runtime"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/bitrise-io/bitrise/analytics"
 	"github.com/bitrise-io/bitrise/bitrise"
 	"github.com/bitrise-io/bitrise/configs"
-	"github.com/bitrise-io/bitrise/exitcode"
 	"github.com/bitrise-io/bitrise/models"
 	"github.com/bitrise-io/bitrise/plugins"
 	"github.com/bitrise-io/bitrise/toolkits"
 	"github.com/bitrise-io/bitrise/tools"
-	"github.com/bitrise-io/bitrise/tools/timeoutcmd"
 	"github.com/bitrise-io/envman/env"
 	envmanModels "github.com/bitrise-io/envman/models"
 	"github.com/bitrise-io/go-utils/colorstring"
-	"github.com/bitrise-io/go-utils/command"
-	"github.com/bitrise-io/go-utils/command/git"
 	"github.com/bitrise-io/go-utils/fileutil"
 	"github.com/bitrise-io/go-utils/pathutil"
 	"github.com/bitrise-io/go-utils/pointers"
@@ -1078,227 +1072,4 @@ func addTestMetadata(testDirPath string, testResultStepInfo models.TestResultSte
 		}
 	}
 	return nil
-}
-
-type BuildRunResultRegisterer struct {
-	tracker analytics.Tracker
-}
-
-func NewBuildRunResultRegisterer(tracker analytics.Tracker) BuildRunResultRegisterer {
-	return BuildRunResultRegisterer{tracker: tracker}
-}
-
-func (r BuildRunResultRegisterer) registerStepRunResults(
-	buildRunResults models.BuildRunResultsModel,
-	stepStartTime time.Time,
-	step stepmanModels.StepModel,
-	stepInfoPtr stepmanModels.StepInfoModel,
-	stepIdxPtr int,
-	runIf string,
-	resultCode int,
-	exitCode int,
-	err error,
-	isLastStep bool,
-	redactedStepInputs map[string]string,
-	properties coreanalytics.Properties) {
-
-	timeout, noOutputTimeout := time.Duration(-1), time.Duration(-1)
-	if resultCode == models.StepRunStatusCodeFailed {
-		// Forward the status of a Step or a wrapped bitrise process.
-		switch exitCode {
-		case exitcode.CLIAbortedWithCustomTimeout:
-			resultCode = models.StepRunStatusAbortedWithCustomTimeout
-		case exitcode.CLIAbortedWithNoOutputTimeout:
-			resultCode = models.StepRunStatusAbortedWithNoOutputTimeout
-		}
-
-		var timeoutErr timeoutcmd.TimeoutError
-		if ok := errors.As(err, &timeoutErr); ok {
-			resultCode = models.StepRunStatusAbortedWithCustomTimeout
-			timeout = timeoutErr.Timeout
-		}
-
-		var noOutputTimeoutErr timeoutcmd.NoOutputTimeoutError
-		if ok := errors.As(err, &noOutputTimeoutErr); ok {
-			resultCode = models.StepRunStatusAbortedWithNoOutputTimeout
-			noOutputTimeout = noOutputTimeoutErr.Timeout
-		}
-	}
-
-	stepInfoCopy := stepmanModels.StepInfoModel{
-		Library:         stepInfoPtr.Library,
-		ID:              stepInfoPtr.ID,
-		Version:         stepInfoPtr.Version,
-		OriginalVersion: stepInfoPtr.OriginalVersion,
-		LatestVersion:   stepInfoPtr.LatestVersion,
-		GroupInfo:       stepInfoPtr.GroupInfo,
-		Step:            stepInfoPtr.Step,
-		DefinitionPth:   stepInfoPtr.DefinitionPth,
-	}
-
-	// Print step preparation errors before the step header box,
-	// other errors are printed within the step box.
-	if resultCode == models.StepRunStatusCodePreparationFailed && err != nil {
-		log.Errorf("Preparing Step (%s) failed: %s", pointers.StringWithDefault(stepInfoCopy.Step.Title, "missing title"), err)
-	}
-	bitrise.PrintRunningStepHeader(stepInfoPtr, step, stepIdxPtr)
-
-	errStr := ""
-	if err != nil {
-		errStr = err.Error()
-	}
-
-	stepResults := models.StepRunResultsModel{
-		StepInfo:   stepInfoCopy,
-		StepInputs: redactedStepInputs,
-		Status:     resultCode,
-		Idx:        buildRunResults.ResultsCount(),
-		RunTime:    time.Since(stepStartTime),
-		ErrorStr:   errStr,
-		ExitCode:   exitCode,
-		StartTime:  stepStartTime,
-	}
-
-	r.tracker.SendStepFinishedEvent(properties, analytics.StepResult{
-		Info:            prepareAnalyticsStepInfo(step, stepInfoPtr),
-		Status:          resultCode,
-		ErrorMessage:    errStr,
-		Timeout:         timeout,
-		NoOutputTimeout: noOutputTimeout,
-	})
-
-	switch resultCode {
-	case models.StepRunStatusCodeSuccess:
-		buildRunResults.SuccessSteps = append(buildRunResults.SuccessSteps, stepResults)
-	case models.StepRunStatusCodePreparationFailed:
-		buildRunResults.FailedSteps = append(buildRunResults.FailedSteps, stepResults)
-	case models.StepRunStatusCodeFailed:
-		log.Errorf("Step (%s) failed: %s", pointers.StringWithDefault(stepInfoCopy.Step.Title, "missing title"), err)
-		buildRunResults.FailedSteps = append(buildRunResults.FailedSteps, stepResults)
-	case models.StepRunStatusAbortedWithCustomTimeout, models.StepRunStatusAbortedWithNoOutputTimeout:
-		log.Errorf("Step (%s) aborted: %s", pointers.StringWithDefault(stepInfoCopy.Step.Title, "missing title"), err)
-		buildRunResults.FailedSteps = append(buildRunResults.FailedSteps, stepResults)
-	case models.StepRunStatusCodeFailedSkippable:
-		if err != nil {
-			log.Warnf("Step (%s) failed, but was marked as skippable: %s", pointers.StringWithDefault(stepInfoCopy.Step.Title, "missing title"), err)
-		} else {
-			log.Warnf("Step (%s) failed, but was marked as skippable", pointers.StringWithDefault(stepInfoCopy.Step.Title, "missing title"))
-		}
-		buildRunResults.FailedSkippableSteps = append(buildRunResults.FailedSkippableSteps, stepResults)
-	case models.StepRunStatusCodeSkipped:
-		log.Warnf("A previous step failed, and this step (%s) was not marked as IsAlwaysRun, skipped", pointers.StringWithDefault(stepInfoCopy.Step.Title, "missing title"))
-		buildRunResults.SkippedSteps = append(buildRunResults.SkippedSteps, stepResults)
-	case models.StepRunStatusCodeSkippedWithRunIf:
-		log.Warn("The step's (" + pointers.StringWithDefault(stepInfoCopy.Step.Title, "missing title") + ") Run-If expression evaluated to false - skipping")
-		if runIf != "" {
-			log.Info("The Run-If expression was: ", colorstring.Blue(runIf))
-		}
-		buildRunResults.SkippedSteps = append(buildRunResults.SkippedSteps, stepResults)
-	default:
-		log.Error("Unknown result code")
-		return
-	}
-
-	bitrise.PrintRunningStepFooter(stepResults, isLastStep)
-}
-
-type StepActivator struct {
-}
-
-func NewStepActivator() StepActivator {
-	return StepActivator{}
-}
-
-func (a StepActivator) activateStep(
-	stepIDData models.StepIDData,
-	buildRunResults models.BuildRunResultsModel,
-	stepDir string,
-	stepYMLPth string,
-	workflowStep stepmanModels.StepModel,
-	stepInfoPtr stepmanModels.StepInfoModel,
-) (string, error) {
-	var origStepYMLPth string
-
-	if stepIDData.SteplibSource == "path" {
-		log.Debugf("[BITRISE_CLI] - Local step found: (path:%s)", stepIDData.IDorURI)
-		stepAbsLocalPth, err := pathutil.AbsPath(stepIDData.IDorURI)
-		if err != nil {
-			return "", err
-		}
-
-		log.Debugln("stepAbsLocalPth:", stepAbsLocalPth, "|stepDir:", stepDir)
-
-		origStepYMLPth = filepath.Join(stepAbsLocalPth, "step.yml")
-		if err := command.CopyFile(origStepYMLPth, stepYMLPth); err != nil {
-			return "", err
-		}
-
-		if err := command.CopyDir(stepAbsLocalPth, stepDir, true); err != nil {
-			return "", err
-		}
-	} else if stepIDData.SteplibSource == "git" {
-		log.Debugf("[BITRISE_CLI] - Remote step, with direct git uri: (uri:%s) (tag-or-branch:%s)", stepIDData.IDorURI, stepIDData.Version)
-		repo, err := git.New(stepDir)
-		if err != nil {
-			return "", err
-		}
-		cloneCmd := repo.CloneTagOrBranch(stepIDData.IDorURI, stepIDData.Version)
-		out, err := cloneCmd.RunAndReturnTrimmedCombinedOutput()
-		if err != nil {
-			if strings.HasPrefix(stepIDData.IDorURI, "git@") {
-				log.Warnf(`Note: if the step's repository is an open source one,
-you should probably use a "https://..." git clone URL,
-instead of the "git@..." git clone URL which usually requires authentication
-even if the repository is open source!`)
-			}
-			var exitErr *exec.ExitError
-			if errors.As(err, &exitErr) {
-				return "", fmt.Errorf("command failed with exit status %d (%s): %w", exitErr.ExitCode(), cloneCmd.PrintableCommandArgs(), errors.New(out))
-			}
-			return "", err
-		}
-
-		if err := command.CopyFile(filepath.Join(stepDir, "step.yml"), stepYMLPth); err != nil {
-			return "", err
-		}
-	} else if stepIDData.SteplibSource == "_" {
-		log.Debugf("[BITRISE_CLI] - Steplib independent step, with direct git uri: (uri:%s) (tag-or-branch:%s)", stepIDData.IDorURI, stepIDData.Version)
-
-		// Steplib independent steps are completly defined in workflow
-		stepYMLPth = ""
-		if err := workflowStep.FillMissingDefaults(); err != nil {
-			return "", err
-		}
-
-		repo, err := git.New(stepDir)
-		if err != nil {
-			return "", err
-		}
-		if err := repo.CloneTagOrBranch(stepIDData.IDorURI, stepIDData.Version).Run(); err != nil {
-			return "", err
-		}
-	} else if stepIDData.SteplibSource != "" {
-		isUpdated := buildRunResults.IsStepLibUpdated(stepIDData.SteplibSource)
-		stepInfo, didUpdate, err := activateStepLibStep(stepIDData, stepDir, stepYMLPth, isUpdated)
-		if didUpdate {
-			buildRunResults.StepmanUpdates[stepIDData.SteplibSource]++
-		}
-
-		stepInfoPtr.ID = stepInfo.ID
-		if stepInfoPtr.Step.Title == nil || *stepInfoPtr.Step.Title == "" {
-			stepInfoPtr.Step.Title = pointers.NewStringPtr(stepInfo.ID)
-		}
-		stepInfoPtr.Version = stepInfo.Version
-		stepInfoPtr.LatestVersion = stepInfo.LatestVersion
-		stepInfoPtr.OriginalVersion = stepInfo.OriginalVersion
-		stepInfoPtr.GroupInfo = stepInfo.GroupInfo
-
-		if err != nil {
-			return "", err
-		}
-	} else {
-		return "", fmt.Errorf("Invalid stepIDData: No SteplibSource or LocalPath defined (%v)", stepIDData)
-	}
-
-	return origStepYMLPth, nil
 }
