+++ cli/step_activator.go
@@ -1,0 +1,118 @@
+package cli
+
+import (
+	"errors"
+	"fmt"
+	"os/exec"
+	"path/filepath"
+	"strings"
+
+	"github.com/bitrise-io/bitrise/models"
+	"github.com/bitrise-io/go-utils/command"
+	"github.com/bitrise-io/go-utils/command/git"
+	"github.com/bitrise-io/go-utils/pathutil"
+	"github.com/bitrise-io/go-utils/pointers"
+	stepmanModels "github.com/bitrise-io/stepman/models"
+	log "github.com/sirupsen/logrus"
+)
+
+type StepActivator struct {
+}
+
+func NewStepActivator() StepActivator {
+	return StepActivator{}
+}
+
+func (a StepActivator) activateStep(
+	stepIDData models.StepIDData,
+	buildRunResults models.BuildRunResultsModel,
+	stepDir string,
+	stepYMLPth string,
+	workflowStep stepmanModels.StepModel,
+	stepInfoPtr stepmanModels.StepInfoModel,
+) (string, error) {
+	var origStepYMLPth string
+
+	if stepIDData.SteplibSource == "path" {
+		log.Debugf("[BITRISE_CLI] - Local step found: (path:%s)", stepIDData.IDorURI)
+		stepAbsLocalPth, err := pathutil.AbsPath(stepIDData.IDorURI)
+		if err != nil {
+			return "", err
+		}
+
+		log.Debugln("stepAbsLocalPth:", stepAbsLocalPth, "|stepDir:", stepDir)
+
+		origStepYMLPth = filepath.Join(stepAbsLocalPth, "step.yml")
+		if err := command.CopyFile(origStepYMLPth, stepYMLPth); err != nil {
+			return "", err
+		}
+
+		if err := command.CopyDir(stepAbsLocalPth, stepDir, true); err != nil {
+			return "", err
+		}
+	} else if stepIDData.SteplibSource == "git" {
+		log.Debugf("[BITRISE_CLI] - Remote step, with direct git uri: (uri:%s) (tag-or-branch:%s)", stepIDData.IDorURI, stepIDData.Version)
+		repo, err := git.New(stepDir)
+		if err != nil {
+			return "", err
+		}
+		cloneCmd := repo.CloneTagOrBranch(stepIDData.IDorURI, stepIDData.Version)
+		out, err := cloneCmd.RunAndReturnTrimmedCombinedOutput()
+		if err != nil {
+			if strings.HasPrefix(stepIDData.IDorURI, "git@") {
+				log.Warnf(`Note: if the step's repository is an open source one,
+you should probably use a "https://..." git clone URL,
+instead of the "git@..." git clone URL which usually requires authentication
+even if the repository is open source!`)
+			}
+			var exitErr *exec.ExitError
+			if errors.As(err, &exitErr) {
+				return "", fmt.Errorf("command failed with exit status %d (%s): %w", exitErr.ExitCode(), cloneCmd.PrintableCommandArgs(), errors.New(out))
+			}
+			return "", err
+		}
+
+		if err := command.CopyFile(filepath.Join(stepDir, "step.yml"), stepYMLPth); err != nil {
+			return "", err
+		}
+	} else if stepIDData.SteplibSource == "_" {
+		log.Debugf("[BITRISE_CLI] - Steplib independent step, with direct git uri: (uri:%s) (tag-or-branch:%s)", stepIDData.IDorURI, stepIDData.Version)
+
+		// Steplib independent steps are completly defined in workflow
+		stepYMLPth = ""
+		if err := workflowStep.FillMissingDefaults(); err != nil {
+			return "", err
+		}
+
+		repo, err := git.New(stepDir)
+		if err != nil {
+			return "", err
+		}
+		if err := repo.CloneTagOrBranch(stepIDData.IDorURI, stepIDData.Version).Run(); err != nil {
+			return "", err
+		}
+	} else if stepIDData.SteplibSource != "" {
+		isUpdated := buildRunResults.IsStepLibUpdated(stepIDData.SteplibSource)
+		stepInfo, didUpdate, err := activateStepLibStep(stepIDData, stepDir, stepYMLPth, isUpdated)
+		if didUpdate {
+			buildRunResults.StepmanUpdates[stepIDData.SteplibSource]++
+		}
+
+		stepInfoPtr.ID = stepInfo.ID
+		if stepInfoPtr.Step.Title == nil || *stepInfoPtr.Step.Title == "" {
+			stepInfoPtr.Step.Title = pointers.NewStringPtr(stepInfo.ID)
+		}
+		stepInfoPtr.Version = stepInfo.Version
+		stepInfoPtr.LatestVersion = stepInfo.LatestVersion
+		stepInfoPtr.OriginalVersion = stepInfo.OriginalVersion
+		stepInfoPtr.GroupInfo = stepInfo.GroupInfo
+
+		if err != nil {
+			return "", err
+		}
+	} else {
+		return "", fmt.Errorf("Invalid stepIDData: No SteplibSource or LocalPath defined (%v)", stepIDData)
+	}
+
+	return origStepYMLPth, nil
+}
